moved code, bug fixes
