+++ src/duHast/Data/Objects/data_sheet.py
@@ -39,8 +39,8 @@ import json
 from duHast.Data.Objects import data_base
 from duHast.Data.Objects.Properties import data_type_properties
 from duHast.Data.Objects.Properties import data_instance_properties
-from duHast.Data.Objects.Properties.Geometry.geometry_bounding_box import (
-    DataBoundingBox,
+from duHast.Data.Objects.Properties.Geometry.geometry_bounding_box_2 import (
+    DataBoundingBox2,
 )
 
 from duHast.Data.Objects.Properties.data_property_names import DataPropertyNames
@@ -65,7 +65,7 @@ class DataSheet(data_base.DataBase):
         self.instance_properties = data_instance_properties.DataInstanceProperties()
         self.type_properties = data_type_properties.DataTypeProperties()
         self.view_ports = []
-        self.bounding_box = DataBoundingBox()
+        self.bounding_box = DataBoundingBox2()
 
         # check if any data was past in with constructor!
         if j != None and len(j) > 0:
@@ -96,7 +96,7 @@ class DataSheet(data_base.DataBase):
                 self.type_properties = data_type_properties.DataTypeProperties(
                     j.get(data_type_properties.DataTypeProperties.data_type, {})
                 )
-                self.bounding_box = DataBoundingBox(
+                self.bounding_box = DataBoundingBox2(
                     j.get(DataPropertyNames.BOUNDING_BOX.value, {})
                 )
