+++ src/duHast/Data/Objects/data_view_schedule.py
@@ -37,8 +37,8 @@ import json
 from duHast.Data.Objects.data_view_base import DataViewBase
 
 from duHast.Data.Objects.Properties.data_property_names import DataPropertyNames
-from duHast.Data.Objects.Properties.Geometry.geometry_bounding_box import (
-    DataBoundingBox,
+from duHast.Data.Objects.Properties.Geometry.geometry_bounding_box_2 import (
+    DataBoundingBox2,
 )
 from duHast.Data.Objects.Properties.data_schedule_segement import DataScheduleSegment
 
@@ -60,7 +60,7 @@ class DataViewSchedule(DataViewBase):
         )
 
         # set default values
-        self.bounding_box = DataBoundingBox()
+        self.bounding_box = DataBoundingBox2()
         self.total_number_of_rows = 0
         self.segments = []
 
@@ -82,7 +82,7 @@ class DataViewSchedule(DataViewBase):
 
             # attempt to populate from json
             try:
-                self.bounding_box = DataBoundingBox(
+                self.bounding_box = DataBoundingBox2(
                     j.get(DataPropertyNames.BOUNDING_BOX.value, {})
                 )
                 self.total_number_of_rows = j.get(DataPropertyNames.TOTAL_NUMBER_OF_ROWS.value, self.total_number_of_rows)
