+++ src/duHast/Revit/Rooms/Export/to_data_room.py
@@ -34,7 +34,7 @@ from duHast.Revit.Common import (
     phases as rPhase,
 )
 from duHast.Data.Objects import data_room as dRoom
-from duHast.Data.Objects.Properties.Geometry import from_revit_conversion as rGeo
+from duHast.Revit.Common.Geometry import to_data_conversion as rGeo
 from duHast.Revit.Rooms.rooms import get_all_rooms
 from duHast.Revit.Rooms.Geometry.geometry import get_2d_points_from_revit_room
 from duHast.Revit.Exports.export_data import (
